Extracting ES-specific code into adapter.
